- merge code nad resolve conflict - login data changed with UserData for common model - implemented actions for Info icon tap event - implementation for edit icon tap (WIP) - based on DB response evaluate and change variables and type in model - used lazyDatabase for use of synchronous database - implemented 7 db queries and manage null to get Specification data - async Task tasks added to purchase order items (WIP) - Exceptions custom dialog for showing exceptions(Partial done)

+++ lib/models/partner_item_sku_inspections.dart
@@ -1,9 +1,9 @@
 import 'package:pverify/services/database/column_names.dart';
 
 class PartnerItemSKUInspections {
-  double? partnerId;
+  int? partnerId;
   String? itemSKU;
-  double? inspectionId;
+  int? inspectionId;
   String? lotNo;
   String? packDate;
   String? lotSize;
@@ -51,9 +51,9 @@ class PartnerItemSKUInspections {
 
   // copyWith
   PartnerItemSKUInspections copyWith({
-    double? partnerId,
+    int? partnerId,
     String? itemSKU,
-    double? inspectionId,
+    int? inspectionId,
     String? lotNo,
     String? packDate,
     String? lotSize,
